Move HPC test to new test_runner.py file.
